+++ package.json
@@ -18,8 +18,6 @@
     "react-router-dom": "^6.4.0",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
-    "sockjs-client": "^1.6.1",
-    "stompjs": "^2.3.3",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.4.2",
     "sweetalert2": "^11.4.33",
