+++ src/components/mypage/MyPageContainer.jsx
@@ -5,7 +5,7 @@ const MyPageContainer = ({ children }) => {
   return <Base>{children}</Base>;
 };
 
-const Base = tw.div`
+const Base = tw.section`
 bg-mGray h-full md:min-h-[26rem] rounded-xl
 `;
