+++ src/pages/Mypage/MyChallenge.js
@@ -1,12 +1,7 @@
-import MyPageContainer from "components/mypage/MyPageContainer";
 import MyChallengeSection from "components/mypage/challenge";
 
 const MyChallenge = () => {
-  return (
-    <MyPageContainer>
-      <MyChallengeSection />
-    </MyPageContainer>
-  );
+  return <MyChallengeSection />;
 };
 
 export default MyChallenge;
