+++ src/pages/Mypage/MyOneLineReviews.js
@@ -1,12 +1,7 @@
-import MyPageContainer from "components/mypage/MyPageContainer";
 import MyOneLineReviewList from "components/mypage/oneLineReview/MyOneLineReviewList";
 
 const MyOneLineReviews = () => {
-  return (
-    <MyPageContainer>
-      <MyOneLineReviewList />
-    </MyPageContainer>
-  );
+  return <MyOneLineReviewList />;
 };
 
 export default MyOneLineReviews;
