+++ src/pages/Mypage/MyPosts.js
@@ -1,13 +1,7 @@
-import React from "react";
 import MyPostsList from "components/mypage/posts/MyPostsList";
-import MyPageContainer from "components/mypage/MyPageContainer";
 
 const MyPosts = () => {
-  return (
-    <MyPageContainer>
-      <MyPostsList />
-    </MyPageContainer>
-  );
+  return <MyPostsList />;
 };
 
 export default MyPosts;
